Article Reading reworked, Header news readable

+++ src/main/java/net/shvdy/nutrition_tracker/model/dao/resultset_mapper/ArticleMapper.java
@@ -25,7 +25,7 @@ public class ArticleMapper implements  ResultSetMapper<List<Article>> {
 
 	private Article extractArticleFromResultSet(ResultSet resultSet) throws SQLException {
 		return Article.builder()
-				.articleId(resultSet.getLong("article_id"))
+				.articleId(resultSet.getInt("article_id"))
 				.authorId(resultSet.getLong("article_id"))
 				.date(resultSet.getString("date_created"))
 				.authorFirstName(resultSet.getString("first_name"))
